+++ scalr/data_source_workspace_test.go
@@ -23,9 +23,11 @@ func TestAccScalrWorkspaceDataSource_basic(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"data.scalr_workspace.test", "auto_apply", "true"),
 					resource.TestCheckResourceAttr(
-						"data.scalr_workspace.test", "terraform_version", "0.12.19"),
+						"data.scalr_workspace.test", "terraform_version", "1.1.9"),
 					resource.TestCheckResourceAttr(
 						"data.scalr_workspace.test", "working_directory", "terraform/test"),
+					resource.TestCheckResourceAttr(
+						"data.scalr_workspace.test", "auto_queue_runs", "skip_first"),
 					resource.TestCheckResourceAttrSet("data.scalr_workspace.test", "environment_id"),
 					resource.TestCheckResourceAttrSet("data.scalr_workspace.test", "has_resources"),
 					resource.TestCheckResourceAttrSet("data.scalr_workspace.test", "created_by.0.full_name"),
@@ -59,7 +61,7 @@ resource scalr_workspace test {
   name                  = "workspace-test-%[1]d"
   environment_id 		= scalr_environment.test.id
   auto_apply            = true
-  terraform_version     = "0.12.19"
+  terraform_version     = "1.1.9"
   working_directory     = "terraform/test"
   hooks {
     pre_init   = "./scripts/pre-init.sh"