+++ scalr/resource_scalr_workspace_test.go
@@ -34,7 +34,7 @@ func TestAccScalrWorkspace_basic(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "operations", "true"),
 					resource.TestCheckResourceAttr(
-						"scalr_workspace.test", "auto_queue_runs", "true"),
+						"scalr_workspace.test", "auto_queue_runs", string(scalr.AutoQueueRunsModeAlways)),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "execution_mode", string(scalr.WorkspaceExecutionModeRemote)),
 					resource.TestCheckResourceAttr(
@@ -103,7 +103,7 @@ func TestAccScalrWorkspace_monorepo(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "operations", "true"),
 					resource.TestCheckResourceAttr(
-						"scalr_workspace.test", "auto_queue_runs", "false"),
+						"scalr_workspace.test", "auto_queue_runs", string(scalr.AutoQueueRunsModeNever)),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "execution_mode", string(scalr.WorkspaceExecutionModeRemote)),
 					resource.TestCheckResourceAttr(
@@ -139,6 +139,8 @@ func TestAccScalrWorkspace_renamed(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "execution_mode", string(scalr.WorkspaceExecutionModeRemote)),
 					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "auto_queue_runs", string(scalr.AutoQueueRunsModeAlways)),
+					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "working_directory", ""),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "hooks.0.pre_init", "./scripts/pre-init.sh"),
@@ -169,6 +171,8 @@ func TestAccScalrWorkspace_renamed(t *testing.T) {
 						"scalr_workspace.test", "operations", "true"),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "execution_mode", string(scalr.WorkspaceExecutionModeRemote)),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "auto_queue_runs", string(scalr.AutoQueueRunsModeAlways)),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "working_directory", ""),
 					resource.TestCheckResourceAttr(
@@ -205,6 +209,8 @@ func TestAccScalrWorkspace_update(t *testing.T) {
 					resource.TestCheckResourceAttr("scalr_workspace.test", "operations", "true"),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "execution_mode", string(scalr.WorkspaceExecutionModeRemote)),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "auto_queue_runs", string(scalr.AutoQueueRunsModeAlways)),
 					resource.TestCheckResourceAttr("scalr_workspace.test", "working_directory", ""),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "run_operation_timeout", "18"),
@@ -238,7 +244,7 @@ func TestAccScalrWorkspace_update(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "execution_mode", string(scalr.WorkspaceExecutionModeLocal)),
 					resource.TestCheckResourceAttr(
-						"scalr_workspace.test", "terraform_version", "0.12.19"),
+						"scalr_workspace.test", "terraform_version", "1.1.9"),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "working_directory", "terraform/test"),
 					resource.TestCheckResourceAttr(
@@ -275,7 +281,7 @@ func TestAccScalrWorkspace_update(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "execution_mode", string(scalr.WorkspaceExecutionModeLocal)),
 					resource.TestCheckResourceAttr(
-						"scalr_workspace.test", "terraform_version", "0.12.19"),
+						"scalr_workspace.test", "terraform_version", "1.1.9"),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "working_directory", "terraform/test"),
 					resource.TestCheckResourceAttr(
@@ -488,7 +494,7 @@ func testAccCheckScalrWorkspaceAttributesUpdated(
 			return fmt.Errorf("Bad execution mode: %s", workspace.ExecutionMode)
 		}
 
-		if workspace.TerraformVersion != "0.12.19" {
+		if workspace.TerraformVersion != "1.1.9" {
 			return fmt.Errorf("Bad Terraform version: %s", workspace.TerraformVersion)
 		}
 
@@ -661,7 +667,7 @@ resource scalr_workspace test {
   auto_apply             = true
   run_operation_timeout = 18
   var_files      = ["test1.tfvars", "test2.tfvars"]
-  auto_queue_runs = true
+  auto_queue_runs = "always"
   hooks {
     pre_init   = "./scripts/pre-init.sh"
     pre_plan   = "./scripts/pre-plan.sh"
@@ -678,7 +684,7 @@ resource "scalr_workspace" "test" {
   name                  = "workspace-monorepo"
   environment_id 		= scalr_environment.test.id
   working_directory     = "/db"
-  auto_queue_runs        = false
+  auto_queue_runs       = "never"
 }`)
 }
 
@@ -688,8 +694,8 @@ resource "scalr_workspace" "test" {
   name                   = "renamed-out-of-band"
   environment_id         = scalr_environment.test.id
   auto_apply             = true
-  auto_queue_runs        = true
   run_operation_timeout = 18
+  auto_queue_runs       = "always"
   hooks {
     pre_init   = "./scripts/pre-init.sh"
     pre_plan   = "./scripts/pre-plan.sh"
@@ -708,7 +714,7 @@ resource "scalr_workspace" "test" {
   environment_id 		= scalr_environment.test.id
   auto_apply            = false
   execution_mode        = "%s"
-  terraform_version     = "0.12.19"
+  terraform_version     = "1.1.9"
   working_directory     = "terraform/test"
   run_operation_timeout = 200
   var_files             = ["test1updated.tfvars", "test2updated.tfvars"]
@@ -731,7 +737,7 @@ resource "scalr_workspace" "test" {
   environment_id 		= scalr_environment.test.id
   auto_apply            = false
   execution_mode        = "%s"
-  terraform_version     = "0.12.19"
+  terraform_version     = "1.1.9"
   working_directory     = "terraform/test"
   vcs_repo {
    identifier = "TestRepo/local"
@@ -749,7 +755,7 @@ resource "scalr_workspace" "test" {
   environment_id 		= scalr_environment.test.id
   auto_apply            = false
   execution_mode        = "%s"
-  terraform_version     = "0.12.19"
+  terraform_version     = "1.1.9"
   working_directory     = "terraform/test"
   vcs_provider_id	    = "test_provider_id"
 }`, scalr.WorkspaceExecutionModeLocal),
@@ -764,7 +770,7 @@ resource "scalr_workspace" "test" {
   environment_id 		= scalr_environment.test.id
   auto_apply            = false
   execution_mode        = "%s"
-  terraform_version     = "0.12.19"
+  terraform_version     = "1.1.9"
   working_directory     = "terraform/test"
 }`, scalr.WorkspaceExecutionModeLocal),
 	)
@@ -778,7 +784,7 @@ resource "scalr_workspace" "test" {
   environment_id 		= scalr_environment.test.id
   auto_apply            = false
   execution_mode        = "%s"
-  terraform_version     = "0.12.19"
+  terraform_version     = "1.1.9"
   working_directory     = ""
   hooks {
     pre_init   = "./scripts/pre-init_updated.sh"